moved edit functions to seperate reducer
