+++ src/actions/index.js
@@ -5,3 +5,4 @@ export * from './navigation'
 export * from './camera'
 export * from './creation'
 export * from './identification'
+export * from './edit'
